# .archive/reference-snapshot/sih/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Cargo": {
        "properties": {
          "quantity_t": {
            "exclusiveMinimum": 0.0,
            "maximum": 2000000.0,
            "title": "Quantity T",
            "type": "number"
          },
          "type": {
            "enum": [
              "thermal_coal",
              "coking_coal",
              "iron_ore"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "quantity_t"
        ],
        "title": "Cargo",
        "type": "object"
      },
...
      "CongestionOut": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "expected_wait_hours_p50": {
            "title": "Expected Wait Hours P50",
            "type": "number"
          },
          "p90": {
            "title": "P90",
            "type": "number"
          },
          "sample_size": {
            "title": "Sample Size",
            "type": "integer"
          },
          "strata": {
            "title": "Strata",
            "type": "string"
          }
        },
        "required": [
          "expected_wait_hours_p50",
          "p90",
          "strata",
          "sample_size",
          "as_of"
        ],
        "title": "CongestionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortDetail": {
        "properties": {
          "approach_channel_depth_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approach Channel Depth M"
          },
          "authority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authority"
          },
          "authority_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authority Type"
          },
          "berths": {
            "items": {
              "$ref": "#/components/schemas/BerthOut"
            },
            "title": "Berths",
            "type": "array"
          },
          "congestion": {
            "$ref": "#/components/schemas/CongestionOut"
          },
          "constraints_as_of": {
            "format": "date",
            "title": "Constraints As Of",
            "type": "string"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "port_id": {
            "title": "Port Id",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date",
            "title": "Retrieved At",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "seasonal_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seasonal Notes"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "tidal_range_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tidal Range M"
          },
          "verification": {
            "title": "Verification",
            "type": "string"
          },
          "vessel_compatibility": {
            "items": {
              "type": "object"
            },
            "title": "Vessel Compatibility",
            "type": "array"
          },
          "weather": {
            "items": {
              "type": "object"
            },
            "title": "Weather",
            "type": "array"
          }
        },
        "required": [
          "port_id",
          "name",
          "country",
          "role",
          "lat",
          "lon",
          "authority",
          "authority_type",
          "approach_channel_depth_m",
          "tidal_range_m",
          "seasonal_notes",
          "berths",
          "congestion",
          "weather",
          "vessel_compatibility",
          "source_url",
          "retrieved_at",
          "verification",
          "constraints_as_of",
          "request_id"
        ],
        "title": "PortDetail",
        "type": "object"
      },
...
      "PortSummary": {
        "properties": {
          "authority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authority"
          },
          "authority_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authority Type"
          },
          "berth_count": {
            "title": "Berth Count",
            "type": "integer"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "max_draft_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Draft M"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "port_id": {
            "title": "Port Id",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "verification": {
            "title": "Verification",
            "type": "string"
          }
        },
        "required": [
          "port_id",
          "name",
          "country",
          "role",
          "lat",
          "lon",
          "authority",
          "authority_type",
          "berth_count",
          "max_draft_m",
          "verification"
        ],
        "title": "PortSummary",
        "type": "object"
      },
      "RecommendVesselRequest": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "cargo": {
            "$ref": "#/components/schemas/Cargo"
          },
          "destination_port": {
            "title": "Destination Port",
            "type": "string"
          },
          "origin_port": {
            "title": "Origin Port",
            "type": "string"
          },
          "required_by": {
            "format": "date",
            "title": "Required By",
            "type": "string"
          }
        },
        "required": [
          "cargo",
          "origin_port",
          "destination_port",
          "required_by"
        ],
        "title": "RecommendVesselRequest",
        "type": "object"
      },
      "RecommendVesselResponse": {
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/VesselOption"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "constraints_as_of": {
            "format": "date",
            "title": "Constraints As Of",
            "type": "string"
          },
          "provenance": {
            "additionalProperties": {
              "enum": [
                "measured",
                "derived",
                "estimated",
                "simulated_demo",
                "expert_set"
              ],
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "recommended": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VesselOption"
              },
              {
                "type": "null"
              }
            ]
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/Rejection"
            },
            "title": "Rejected",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "recommended",
          "alternatives",
          "rejected",
          "constraints_as_of",
          "as_of",
          "assumptions",
          "provenance",
          "request_id"
        ],
        "title": "RecommendVesselResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VesselClassOut": {
        "properties": {
          "cons_ballast_mt_day": {
            "title": "Cons Ballast Mt Day",
            "type": "number"
          },
          "cons_laden_mt_day": {
            "title": "Cons Laden Mt Day",
            "type": "number"
          },
          "constants_t": {
            "title": "Constants T",
            "type": "number"
          },
          "demurrage_usd_day": {
            "title": "Demurrage Usd Day",
            "type": "number"
          },
          "dwt_range": {
            "items": {
              "type": "number"
            },
            "title": "Dwt Range",
            "type": "array"
          },
          "geared": {
            "title": "Geared",
            "type": "boolean"
          },
          "grt": {
            "title": "Grt",
            "type": "number"
          },
          "index_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Code"
          },
          "ref_beam_m": {
            "title": "Ref Beam M",
            "type": "number"
          },
          "ref_draft_m": {
            "title": "Ref Draft M",
            "type": "number"
          },
          "ref_dwt": {
            "title": "Ref Dwt",
            "type": "number"
          },
          "ref_loa_m": {
            "title": "Ref Loa M",
            "type": "number"
          },
          "retrieved_at": {
            "format": "date",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "speed_ballast_kn": {
            "title": "Speed Ballast Kn",
            "type": "number"
          },
          "speed_laden_kn": {
            "title": "Speed Laden Kn",
            "type": "number"
          },
          "tpc": {
            "title": "Tpc",
            "type": "number"
          },
          "vessel_class": {
            "title": "Vessel Class",
            "type": "string"
          }
        },
        "required": [
          "vessel_class",
          "index_code",
          "ref_dwt",
          "ref_draft_m",
          "ref_loa_m",
          "ref_beam_m",
          "tpc",
          "speed_laden_kn",
          "speed_ballast_kn",
          "cons_laden_mt_day",
          "cons_ballast_mt_day",
          "geared",
          "dwt_range",
          "constants_t",
          "demurrage_usd_day",
          "grt",
          "source_url",
          "retrieved_at"
        ],
        "title": "VesselClassOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nDecision-support API for chartering bulk carriers into India's East Coast ports.\n\n**Forecast the market \u2192 understand the constraints \u2192 simulate charter strategies \u2192\nrecommend the best decision.**\n\nThis is not a freight price oracle. It is a constrained decision engine that treats\nthe freight forecast as an uncertain input, not as the answer. Every response carries\n`provenance`, `as_of` and `assumptions`, and every infeasible option comes back with\nthe constraint and the source that ruled it out.\n\nProvenance values: `measured` (from an official published record), `derived` (computed\nby one of our models from measured inputs), `estimated` (a documented judgement where no\npublication exists), `expert_set` (a hand-curated parameter), `simulated_demo` (a\nseeded demo series standing in for a feed that needs credentials).\n",
    "title": "SIH26006 \u2014 Charter Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ports": {
      "get": {
        "operationId": "list_ports_api_v1_ports_get",
        "parameters": [
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(load|discharge|both)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortSummary"
                  },
                  "title": "Response List Ports Api V1 Ports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Port registry",
        "tags": [
          "ports"
        ]
      }
    },
    "/api/v1/ports/{port_id}": {
      "get": {
        "operationId": "get_port_api_v1_ports__port_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "port_id",
            "required": true,
            "schema": {
              "title": "Port Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cargo_type",
            "required": false,
            "schema": {
              "default": "coking_coal",
              "title": "Cargo Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Port constraints, congestion and compatibility",
        "tags": [
          "ports"
        ]
      }
    },
    "/api/v1/ports/{port_id}/compatibility": {
      "get": {
        "operationId": "port_compatibility_api_v1_ports__port_id__compatibility_get",
        "parameters": [
          {
            "in": "path",
            "name": "port_id",
            "required": true,
            "schema": {
              "title": "Port Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "cargo_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cargo Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Full compatibility grid for a port",
        "tags": [
          "ports"
        ]
      }
    },
    "/api/v1/recommend-vessel": {
      "post": {
        "operationId": "recommend_vessel_api_v1_recommend_vessel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendVesselRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendVesselResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feasibility filter, parcel split and class ranking",
        "tags": [
          "vessels"
        ]
      }
    },
...
    "/api/v1/vessels/{vessel_class}": {
      "get": {
        "operationId": "get_vessel_class_api_v1_vessels__vessel_class__get",
        "parameters": [
          {
            "in": "path",
            "name": "vessel_class",
            "required": true,
            "schema": {
              "title": "Vessel Class",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VesselClassOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vessel Class",
        "tags": [
          "vessels"
        ]
      }
    }
  },
...
}